Use floating-point versions of math functions

+++ source/modules/abs.cpp
@@ -7,5 +7,5 @@ Abs::Abs()
 }
 
 void Abs::process(uint32_t fs) {
-    outputs[0].value = fabs(inputs[0].value);
+    outputs[0].value = fabsf(inputs[0].value);
 }
